Cleanup oauth app activation path
We no longer use the oauth app, and the GitHub app flow doesn't require
this code, so it's basically dead.
