+++ spec/controllers/activations_controller_spec.rb
@@ -5,71 +5,34 @@ describe ActivationsController, "#create" do
     it "returns successful response" do
       membership = create(:membership)
       repo = membership.repo
-      activator = double("RepoActivator", activate: true, errors: [])
-      allow(RepoActivator).to receive(:new).and_return(activator)
       stub_sign_in(membership.user)
-      expected_repo_json = RepoSerializer.new(
-        repo,
-        scope_name: :current_user,
-        scope: membership.user,
-      ).to_json
 
       post :create, params: { repo_id: repo.id }, format: :json
 
       expect(response).to have_http_status(:created)
-      expect(response.body).to eq expected_repo_json
-      expect(activator).to have_received(:activate)
-      expect(RepoActivator).to have_received(:new).
-        with(repo: repo, github_token: membership.user.token)
-    end
-  end
-
-  context "when activation fails" do
-    context "due to 403 Forbidden from GitHub" do
-      it "returns error response" do
-        membership = create(:membership)
-        repo = membership.repo
-        error_message = "You must be an admin to add a team membership"
-        activator = double(
-          "RepoActivator",
-          activate: false,
-          errors: [error_message]
-        )
-        allow(RepoActivator).to receive(:new).and_return(activator)
-        stub_sign_in(membership.user)
-
-        post :create, params: { repo_id: repo.id }, format: :json
-
-        response_body = JSON.parse(response.body)
-        expect(response.code).to eq "502"
-        expect(response_body["errors"]).to match_array(error_message)
-        expect(activator).to have_received(:activate)
-        expect(RepoActivator).to have_received(:new).
-          with(repo: repo, github_token: membership.user.token)
-      end
-    end
-
-    it "tracks failed activation" do
-      membership = create(:membership)
-      repo = membership.repo
-      activator = double(
-        "RepoActivator",
-        activate: false,
-        errors: []
+      expect(JSON.parse(response.body)).to eq(
+        "id" => repo.id,
+        "active" => true,
+        "admin" => false,
+        "github_id" => repo.github_id,
+        "name" => repo.name,
+        "price_in_cents" => 0,
+        "private" => false,
+        "stripe_subscription_id" => nil,
+        "owner" => {
+          "id" => repo.owner.id,
+          "created_at" => repo.owner.created_at.iso8601(3),
+          "updated_at" => repo.owner.updated_at.iso8601(3),
+          "github_id" => repo.owner.github_id,
+          "name" => repo.owner.name,
+          "organization" => false,
+          "config_enabled" => false,
+          "config_repo" => nil,
+          "whitelisted" => false,
+          "marketplace_plan_id" => nil,
+          "stripe_subscription_id" => nil,
+        },
       )
-      allow(RepoActivator).to receive(:new).and_return(activator)
-      stub_sign_in(membership.user)
-
-      post :create, params: { repo_id: repo.id }, format: :json
-
-      expect(analytics).to have_tracked("Repo Activation Failed").
-        for_user(membership.user).
-        with(
-          properties: {
-            name: repo.name,
-            private: false
-          }
-        )
     end
   end
 
@@ -78,13 +41,10 @@ describe ActivationsController, "#create" do
       repo = create(:repo, private: true)
       user = create(:user)
       user.repos << repo
-      activator = double("RepoActivator", activate: false)
-      allow(RepoActivator).to receive(:new).and_return(activator)
       stub_sign_in(user)
 
       expect { post :create, params: { repo_id: repo.id }, format: :json }.
         to raise_error(ActivationsController::CannotActivatePaidRepo)
-      expect(activator).not_to have_received(:activate)
     end
   end
 end
